Place undock and refresh list actions in the same ribbon group Change colour of undock button to blue
