+++ Post-Edit Compare/Sdl.Community.PostEdit.Compare.Versions/HTMLReportIntegration/Ribbon/SynchronizationRibbonGroup.cs
@@ -2,17 +2,10 @@
 using Sdl.Desktop.IntegrationApi;
 using Sdl.Desktop.IntegrationApi.Extensions;
 using Sdl.Desktop.IntegrationApi.Extensions.Internal;
-using Sdl.TranslationStudioAutomation.IntegrationApi.Presentation.DefaultLocations;
 using System;
 
 namespace Sdl.Community.PostEdit.Versions.HTMLReportIntegration.Ribbon
 {
-    [RibbonGroup(nameof(SynchronizationRibbonGroup), "Synchronization")]
-    [RibbonGroupLayout(LocationByType = typeof(TranslationStudioDefaultRibbonTabs.HomeRibbonTabLocation))]
-    public class SynchronizationRibbonGroup : AbstractRibbonGroup
-    {
-    }
-
     [Action(nameof(SyncReportProjectOn),
             typeof(ReportViewController),
         Icon = "Sync", Name = "Synchronize",
